Phishion/requirement.txt
aiohttp==3.9.1
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
ansicon==1.89.0
anyio==4.2.0
asttokens==2.4.1
atomicwrites==1.4.1
attrs==23.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses==0.6
dataclasses-json==0.6.3
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
editor==1.6.5
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.1
future==0.18.3
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.16.1
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
gotrue==2.4.2
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==2.10
iniconfig==2.0.0
inquirer==3.2.1
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jinxed==1.2.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.13
langsmith==0.0.83
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==0.28.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pefile==2023.2.7
pillow==10.2.0
pluggy==1.4.0
postgrest==0.16.2
postgrest-py==0.4.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pychrome==0.2.4
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-googlesearch==1.1.1
python-whois==0.8.0
pytz==2024.1
pyuseragents==1.0.5
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
readchar==4.0.5
realtime==1.0.3
realtime-py==0.1.3
reportlab==4.0.9
requests==2.25.1
requests-file==1.5.1
rfc3986==1.5.0
rsa==4.9
runs==1.2.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.1
supafunc==0.4.5
tenacity==8.2.3
threadpoolctl==3.2.0
tldextract==5.1.1
toml==0.10.2
tqdm==4.66.1
traitlets==5.14.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
xmod==1.8.1
yarl==1.9.4
